app/requirements.txt
Flask
python-dotenv
PyJWT
requests
pytest
pydantic
requests
gunicorn
gevent
urllib3
SQLAlchemy
psycopg2-binary
cachetools
uuid